+++ app/client/src/components/editorComponents/ApiResponseView.tsx
@@ -33,7 +33,7 @@ import { ApiResponseHeaders } from "PluginActionEditor/components/PluginActionRe
 interface Props {
   currentActionConfig: Action;
   theme?: EditorTheme;
-  disabled: boolean;
+  isRunDisabled: boolean;
   onRunClick: () => void;
   actionResponse?: ActionResponse;
   isRunning: boolean;
@@ -43,7 +43,7 @@ function ApiResponseView(props: Props) {
   const {
     actionResponse = EMPTY_RESPONSE,
     currentActionConfig,
-    disabled,
+    isRunDisabled = false,
     isRunning,
     theme = EditorTheme.LIGHT,
   } = props;
@@ -99,7 +99,7 @@ function ApiResponseView(props: Props) {
         <ApiResponse
           action={currentActionConfig}
           actionResponse={actionResponse}
-          isRunDisabled={disabled}
+          isRunDisabled={isRunDisabled}
           isRunning={isRunning}
           onRunClick={onRunClick}
           responseTabHeight={responseTabHeight}
@@ -113,7 +113,7 @@ function ApiResponseView(props: Props) {
       panelComponent: (
         <ApiResponseHeaders
           actionResponse={actionResponse}
-          isRunDisabled={disabled}
+          isRunDisabled={isRunDisabled}
           isRunning={isRunning}
           onDebugClick={onDebugClick}
           onRunClick={onRunClick}
