+++ app/client/src/pages/Editor/APIEditor/CommonEditorForm.tsx
@@ -341,7 +341,7 @@ function CommonEditorForm(props: CommonFormPropsWithExtraParams) {
               <ApiResponseView
                 actionResponse={actionResponse}
                 currentActionConfig={currentActionConfig}
-                disabled={!isExecutePermitted}
+                isRunDisabled={blockExecution}
                 isRunning={isRunning}
                 onRunClick={onRunClick}
                 theme={theme}
